Keep loginWindow Centered, Added util.js, cleaned login.js
loginWindow now has a listener (via .on method) which keeps loginWindow centered.

The util.js will be used later as we end up having a need to add random things to the code base. This is basically our junk drawer (let's keep it full but tidy/orderly).

Finally, login.js had junk in it so I removed it.  We can use this later to provide a "Please verify your login" prompt when a customer does something sensitive or they have been idle too long.